+++ src/libtcod/sys_sdl_c.cpp
@@ -520,7 +520,8 @@ void find_resolution(void) {
 	sdl->get_closest_mode(&TCOD_ctx.actual_fullscreen_width,&TCOD_ctx.actual_fullscreen_height);
 }
 
-SDL_Surface* TCOD_sys_create_bitmap_for_console(TCOD_Console* console) {
+SDL_Surface* TCOD_sys_create_bitmap_for_console(const TCOD_Console* console)
+{
   int w = TCOD_console_get_width(console) * TCOD_ctx.font_width;
   int h = TCOD_console_get_height(console) * TCOD_ctx.font_height;
   return TCOD_sys_get_surface(w, h, false);
@@ -553,7 +554,7 @@ static TCOD_color_t blend_fade_color_(
 
 void TCOD_sys_console_to_bitmap(
     SDL_Surface* vbitmap,
-    struct TCOD_Console* console,
+    const struct TCOD_Console* console,
     struct TCOD_Console* cache)
 {
   static SDL_Surface* charmap_backup = NULL;