+++ src/libtcod/txtfield.cpp
@@ -63,8 +63,9 @@ bool TCODText::update(TCOD_key_t key){
 	return TCOD_text_update(data,key) != 0;
 }
 
-void TCODText::render(TCODConsole * con){
-	TCOD_text_render(data,con->data);
+void TCODText::render(TCODConsole * con)
+{
+  TCOD_text_render(data,con->get_data());
 }
 
 const char *TCODText::getText(){